% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afwijkendeMetingen.R
\name{afwijkendeMetingen}
\alias{afwijkendeMetingen}
\title{Geeft de afwijkende metingen uit een gegeven model}
\usage{
afwijkendeMetingen(Basismodel, Rmse_domein, Verschuiving, Rmse_verschuiving)
}
\arguments{
\item{Basismodel}{model per boomsoort}

\item{Rmse.basis}{rmse_domein en rmse_Vlaams}

\item{Afgeleidmodel}{ingeval van verschoven Vlaams model}

\item{Rmse.afgeleid}{ingeval van verschoven Vlaams model}
}
\value{
lijst met afwijkende metingen (> 2,5 * rmse), inclusief vlag uit databank
}
\description{
Berekent afwijkende metingen, dit zijn metingen met een afwijking > 2,5 * rmse, en geeft deze weer volgens dalende afwijking
}

