% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_basis.R
\name{fit.basis}
\alias{fit.basis}
\title{Basismodel fitten op basis van de opgegeven dataset}
\usage{
fit.basis(Data.basis)
}
\arguments{
\item{Data.basis}{dataframe > 50 en min. 6 domeinen}
}
\value{
list met model per boomsoort
}
\description{
Functie die het lineair mixed model fit op basis van de opgegeven dataset.  Idee is dat deze functie gebruikt wordt voor het fitten van het volledige model, maar ook voor het fitten van modellen op basis van een gedeeltelijke dataset (zie rmse.basis)
}
