% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lokaal.R
\name{fit.lokaal}
\alias{fit.lokaal}
\title{Lokaal model fitten op basis van de opgegeven dataset}
\usage{
fit.lokaal(Data.lokaal)
}
\arguments{
\item{Data.lokaal}{dataframe > 50 en < 6 domeinen}
}
\value{
list met model per boomsoort-domein-combinatie
}
\description{
Functie die het lineair (fixed) model fit op basis van de opgegeven dataset.  Idee is dat deze functie gebruikt wordt voor het fitten van het volledige model, maar ook voor het fitten van modellen op basis van een gedeeltelijke dataset (zie rmse.lokaal)
}
