% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoogteschatting_afgeleid.R
\name{hoogteschatting.afgeleid}
\alias{hoogteschatting.afgeleid}
\title{Hoogteschatting op basis van opgegeven afgeleid model}
\usage{
hoogteschatting.afgeleid(Afgeleidmodel, Data.afgeleid)
}
\arguments{
\item{Afgeleidmodel}{verschuiving per boomsoort en domein (verschoven Vlaams model)}

\item{Data.afgeleid}{dataframe 10-50}
}
\value{
dataframe met de meetresultaten en de schattingen van de hoogtes voor het domeinmodel en de Vlaamse model
}
\description{
Functie die de gemiddelde hoogte per omtrekklasse schat voor de domeincurves en Vlaamse curves van het opgegeven afgeleid model.  De teruggegeven dataframe kan gebruikt worden om grafieken te maken of afwijkende metingen te bestuderen.
}

