% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoogteschatting_basis.R
\name{hoogteschatting.basis}
\alias{hoogteschatting.basis}
\title{Hoogteschatting op basis van opgegeven basismodel}
\usage{
hoogteschatting.basis(Soortmodel, Soortdata, Typemodel)
}
\arguments{
\item{Soortmodel}{model voor boomsoort (basis) of boomsoort-domein-combinatie (extra)}

\item{Soortdata}{meetgegevens van boomsoort (basis) of domein-boomsoort-combinatie (extra)}

\item{Typemodel}{'Basis' of 'Extra'?}
}
\value{
dataframe met de meetresultaten en de schattingen van de hoogtes voor het domeinmodel en de Vlaamse model
}
\description{
Functie die de gemiddelde hoogte per omtrekklasse schat voor de domeincurves en Vlaamse curves van het opgegeven basismodel.  De teruggegeven dataframe kan gebruikt worden om grafieken te maken of afwijkende metingen te bestuderen.
}

