% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoercontrole.R
\name{invoercontrole}
\alias{invoercontrole}
\title{Hulpfunctie die ingevoerd dataframe controleert}
\usage{
invoercontrole(Data, Type)
}
\arguments{
\item{Data}{De dataframe die moet gecontroleerd worden}

\item{Type}{Het type controle dat moet gebeuren, verwijzend naar de functie(s) waarvoor het gebruikt wordt.  Mogelijke waarden zijn: 'initiatie', 'fit', 'basismodel', 'lokaalmodel', 'afgeleidmodel' en 'afgeleidedata'}
}
\value{
Een foutmelding of de geteste dataframe
}
\description{
Omdat de controle op de invoer voor meerdere functies gelijkaardig is, maken we een functie aan die deze
}
