% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leesFile.R
\name{leesFile}
\alias{leesFile}
\title{Leest tekst in file van package dhcurve}
\usage{
leesFile(Bestandsnaam)
}
\arguments{
\item{Bestandsnaam}{Naam van het tekstbestandje dat in het package
opgeslagen is, bv. "dbpath"}
}
\value{
Inhoud van het opgegeven tekstbestandje
}
\description{
Idee is om in het package zelf (in de map 'inst') 2 gebruikerspecifieke
tekstbestandjes toe te voegen met het path waar de databank opgeslagen is
(='dbpath') en evt. met het path wat als working directory gebruikt zal
worden (='wdpath').  Deze laatste is facultatief, als deze niet opgegeven
wordt, zal de eerder opgegeven working directory gebruikt worden.  Deze
functie kan in de hoofdfunctie of andere scripts buiten het package gebruikt
worden om deze (vaste) locatie op te roepen, wat als voordeel biedt dat de
link naar de databank bij verplaatsen van de databank niet in elk script
aangepast moet worden (maar enkel in dat ene bestand).
}
