% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maakGrafiek.R
\name{maakGrafiek}
\alias{maakGrafiek}
\title{maakt een grafiek van een slechte curve of model met afwijkende metingen}
\usage{
maakGrafiek(Data)
}
\arguments{
\item{Data}{Dataset van 1 boomsoort-domeincombinatie met alle info die nodig is om een grafiek aan te maken: gemeten waarden, geschatte waarde voor domeinmodel en Vlaams model, aanduiding van afwijkende waarden, buigpunten, te tonen info als RMSE, parameterwaarden, reden waarom het model opgenomen is,...}
}
\value{
grafiek van de boomsoort-domeincombinatie waarin de problemen in rood aangeduid zijn
}
\description{
Deze functie genereert een grafiek met volgende info op basis van de gegeven dataset (= data van 1 domein-boomsoort-combinatie!):
}
\details{
- puntenwolk metingen

- curve Vlaams model

- curve domeinmodel of verschoven Vlaams model (dus domeinspecifieke curve)

- extremen en buigpunten waar aanwezig (met vertikale stippellijn?)

- grenzen van bruikbaar interval

- vermelding van boomsoort, domein, aantal waarnemingen, rmse, evt. parameters A, B en C

- eventueel afwijkende metingen

Idee is dat in de grafieken een afwijkende kleur het probleem aangeeft: afwijkende meting, buigpunt,...  Daarnaast wordt het probleem ook tekstueel weergegeven
}
