% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelparameters_afgeleid.R
\name{modelparameters.afgeleid}
\alias{modelparameters.afgeleid}
\title{Modelparameters berekenen voor opgegeven afgeleid model}
\usage{
modelparameters.afgeleid(Soortmodel)
}
\arguments{
\item{Soortmodel}{model voor boomsoort-domeincombinatie}
}
\value{
dataframe met parameter voor de verschuiving t.o.v. het basismodel
(Ad)
}
\description{
Interne functie die de modelparameters berekent op basis van een opgegeven
afgeleid model.  Ze berekent de parameter voor de verschuiving van het
basismodel.
}
