% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelparameters_basis.R
\name{modelparameters.basis}
\alias{modelparameters.basis}
\title{Modelparameters berekenen voor opgegeven basismodel}
\usage{
modelparameters.basis(Soortmodel)
}
\arguments{
\item{Soortmodel}{model van boomsoort}
}
\value{
dataframe met parameters voor domeinmodel (Ad, Bd en Cd) en parameters voor Vlaams model (Avl, Bvl en Cvl)
}
\description{
Functie die de modelparameters berekent op basis van een opgegeven basismodel.  Ze berekent zowel de parameters voor het domeinmodel als voor het Vlaams model, en geeft ook de grenzen van het bruikbaar interval
}

