% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelparameters_extra.R
\name{modelparameters.extra}
\alias{modelparameters.extra}
\title{Modelparameters berekenen voor opgegeven extra model}
\usage{
modelparameters.extra(Extramodellen, Data)
}
\arguments{
\item{Extramodellen}{model per boomsoort-domein-combinatie}

\item{Data}{meetgegevens (enkel nodig voor model per boomsoort-domein-combinatie)}
}
\value{
dataframe met parameters voor domeinmodel (Ad, Bd en Cd)
}
\description{
Functie die de modelparameters berekent op basis van een opgegeven extra model.  Ze berekent de parameters voor het domeinmodel en geeft de grenzen van het bruikbaar interval
}

