% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelparameters_lokaal.R
\name{modelparameters.lokaal}
\alias{modelparameters.lokaal}
\title{Modelparameters berekenen voor opgegeven lokaal model}
\usage{
modelparameters.lokaal(Soortmodel)
}
\arguments{
\item{Soortmodel}{model voor boomsoort-domein-combinatie}
}
\value{
dataframe met parameters voor domeinmodel (\code{Ad}, \code{Bd} en \code{Cd})
}
\description{
Functie die de modelparameters berekent op basis van een opgegeven lokaal
model.  Ze berekent de parameters voor het domeinmodel en geeft de grenzen
van het bruikbaar interval
}
