% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelparameters.R
\name{modelparameters}
\alias{modelparameters}
\title{Modelparameters berekenen voor opgegeven model}
\usage{
modelparameters(Basismodel, Data = NULL)
}
\arguments{
\item{Basismodel}{model per boomsoort of model per boomsoort-domein-combinatie}

\item{Data}{}
}
\value{
dataframe met parameters voor domeinmodel (Ad, Bd en Cd) en ingeval van het basismodel de parameters voor Vlaams model (Avl, Bvl en Cvl)
}
\description{
Functie die de modelparameters berekent op basis van een opgegeven basismodel of extra.  Ze berekent de parameters voor het domeinmodel en ingeval van het basismodel ook voor het Vlaams model, en geeft ook de grenzen van het bruikbaar interval.  (Deze functie verwijst naar de functies modelparameters.basis of modelparameters.extra, afhankelijk van de situatie)
}

