% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputIVANHO.R
\name{outputIVANHO}
\alias{outputIVANHO}
\title{berekent de geschatte hoogtes per omtrekklasse}
\usage{
outputIVANHO(Basismodel, Afgeleidmodel, Lokaalmodel, Data.lokaal)
}
\arguments{
\item{Basismodel}{model per boomsoort}

\item{Afgeleidmodel}{verschuiving per boomsoort en domein (Vlaams model)}

\item{Lokaalmodel}{model per boomsoort-domein-combinatie}

\item{Data.lokaal}{data voor model per boomsoort-domein-combinatie}
}
\value{
dataframe met geschatte hoogtes per domein en per boomsoort met velden:

- ModelID

- DOMEIN_ID: domeincode

- BOS_BHI: domeinnaam

- BoomsoortID

- BMS: boomsoort

- Modeltype ('basismodel'(= eigen model op basis van mixed model) / ‘afgeleid model'(= verschoven Vlaams model, afgeleid van fixed factor uit basismodel) / ‘Vlaams model’(= fixed factor uit basismodel, niet toegevoegd omdat niet relevant) / 'lokaal model'(= apart model voor 1 boomsoort-domein-combinatie) / 'geen model'(= boomsoort-domein-combinatie waarvoor geen model berekend kan worden))

- Omtrek: klassemidden

- OmtrekklassetypeID

- Omtrekklasse

- Hoogte
}
\description{
Berekeningen die leiden tot de gevraagde tabel om in IVANHO te importeren.  Hiervoor worden volgende hulpfuncties aangeroepen:
}
\details{
hoogteschatting.basis, hoogteschatting.afgeleid

Verder moeten parameters A, B en C uit het model gehaald worden (coef) en een aantal berekende gegevens uit de eerste stap toegevoegd worden.
}
