% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultaat.R
\name{resultaat}
\alias{resultaat}
\title{berekent de modelparameters per domein}
\usage{
resultaat(Basismodel, Afgeleidmodel, Extramodellen, Data.extra,
  Data.ontbrekend = NULL)
}
\arguments{
\item{Basismodel}{model per boomsoort}

\item{Afgeleidmodel}{verschuiving per boomsoort en domein (Vlaams model)}

\item{Extramodellen}{model per boomsoort-domein-combinatie}

\item{Data.extra}{data voor model per boomsoort-domein-combinatie}

\item{Data.ontbrekend}{evt. lijst van domeinen < 10 per boomsoort (en dan voor deze domeinen het Vlaams model invullen?)}
}
\value{
dataframe met modellen per domein en per boomsoort met velden:

- ModelID

- DomeinID

- BoomsoortID

- type model ('basismodel'(eigen model op basis van mixed model)/‘eigen model’ (eigen fixed model)/‘afgeleid model' (= verschoven Vlaams model, afgeleid van fixed factor uit basismodel)/‘Vlaams model’ (= fixed factor uit basismodel)/ 'domeinmodel' (= apart model voor 1 boomsoort-domein-combinatie))

- paramaters A,B en C

- bruikbaar interval

- rmse

- aantal metingen waarop model gebaseerd is

- aantal metingen > 0.5 m (dus waarop rmse-berekening gebaseerd is)

evt. kan een tweede dataframe toegevoegd worden met Vlaamse modellen per boomsoort, of deze kan toegevoegd worden aan de vorige dataframe, waarbij DomeinID leeg gelaten wordt of een specifieke waarde ‘Vlaams model’ krijgt
}
\description{
Berekeningen die leiden tot de gevraagde tabel.  Hiervoor worden volgende hulpfuncties aangeroepen:
}
\details{
modelparameters
rsme.basis
fit.afgeleid

Verder moeten parameters A, B en C uit het model gehaald worden (coef) en een aantal berekende gegevens uit de eerste stap toegevoegd worden.
}

