% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_afgeleid.R
\name{rmse.afgeleid}
\alias{rmse.afgeleid}
\title{RMSE berekenen van afgeleid model (verschoven Vlaams model)}
\usage{
rmse.afgeleid(Basismodel, Afgeleidmodel, Data.afgeleid)
}
\arguments{
\item{Basismodel}{model per boomsoort}

\item{Afgeleidmodel}{verschuiving per boomsoort en domein (verschoven Vlaams model)}

\item{Data.afgeleid}{dataframe 10-50}
}
\value{
dataframe met rmse_verschuiving per boomsoort en domein
}
\description{
deze functie berekent de rmse op basis van verschil tussen geschatte domeinwaarde en gemeten waarde (voor omtrekklassen > 0.5 m)
}
\details{
DEZE FUNCTIE IS NIET NODIG, fit.afgeleid BEREKENT METEEN DE RMSE!!!
}

