% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata1domein.R
\name{testdata1domein}
\alias{testdata1domein}
\title{Genereert testdata voor 1 boomsoort-domeincombinatie}
\usage{
testdata1domein(
  nBomen = 100,
  minOmtrek = 20,
  maxOmtrek = 239,
  A = 30,
  B = 15,
  C = 1,
  sd = 2
)
}
\arguments{
\item{nBomen}{Aantal gemeten bomen voor het domein}

\item{minOmtrek}{Minimale omtrek voor de data}

\item{maxOmtrek}{Maximale omtrek voor de data}

\item{A}{Parameter voor het intercept van de functie}

\item{B}{Parameter voor de 1ste-graadsfunctie}

\item{C}{Parameter voor de 2de-graadsfunctie}

\item{sd}{foutenmarge t.o.v. model voor de data}
}
\value{
Dataframe met testdata voor 1 boomsoort-domeincombinatie met velden
C13 en HOOGTE.
}
\description{
Deze functie genereert een testdataset voor 1 domein die voldoet aan de
opgegeven vereisten.
}
