% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdataset.R
\name{testdataset}
\alias{testdataset}
\title{Genereert testdataset}
\usage{
testdataset(
  nBomenDomein = c(200, 100, 500, 100, 100, 100, 50, 20, 8),
  BMS = "testboom",
  IDbms = 1,
  minOmtrek = 20,
  maxOmtrek = 239
)
}
\arguments{
\item{nBomenDomein}{Vector met aantal gemeten bomen per domein}

\item{BMS}{Naam van boomsoort}

\item{IDbms}{Identificatienummer van boomsoort}

\item{minOmtrek}{Minimale omtrek voor de data}

\item{maxOmtrek}{Maximale omtrek voor de data}
}
\value{
Dataframe met testdata voor 1 boomsoort met velden ...
}
\description{
Deze functie genereert een testdataset die voldoet aan de opgegeven
vereisten.
}
