% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatie_afgeleid.R
\name{validatie.afgeleid}
\alias{validatie.afgeleid}
\title{Validatie van het afgeleid model}
\usage{
validatie.afgeleid(
  Basismodel,
  Afgeleidmodel,
  AantalDomHogeRMSE = 20,
  ExtraCurvesRapport = NULL,
  Bestandsnaam = "Validatie_Afgeleid.html",
  TypeRapport = "Dynamisch",
  PathWD = getwd()
)
}
\arguments{
\item{Basismodel}{Model per boomsoort zoals teruggegeven door de functie
\code{fit.basis()}: tibble met de velden \code{BMS} (boomsoort) en \code{Model}
(\code{lme}-object met het gefit mixed model voor die boomsoort)}

\item{Afgeleidmodel}{Model per boomsoort-domeincombinatie zoals teruggegeven
door de functie \code{fit.afgeleid()}: list met 2 tibbles.}

\item{AantalDomHogeRMSE}{Standaard worden de 20 domeinen met de hoogste
\code{RMSE} geselecteerd en voor elk van deze domeinen worden minstens 10
afwijkende metingen geselecteerd. \code{AantalDomHogeRMSE} laat toe om dit aantal
van 20 domeinen aan te passen.}

\item{ExtraCurvesRapport}{Optie om een lijst met extra
boomsoort-domeincombinaties op te geven waardoor de curve getoond moet worden
in het validatierapport.
Deze moeten opgegeven worden als een dataframe met velden \code{DOMEIN_ID} en
\code{BMS}, met benamingen die overeenkomen met deze in de opgegeven dataset.}

\item{Bestandsnaam}{Een naam voor het validatierapport (\code{.html}-bestand) dat
gegenereerd wordt, bestaande uit een string die eindigt op \code{.html}}

\item{TypeRapport}{Default is "Dynamisch", waarbij de figuren in het
\code{.html}-bestand kunnen worden aangepast (meetgegevens weergeven door muis
erover te bewegen (inclusief \code{ID} als deze in de dataset meegegeven is),
items uit legende wegklikken, grafiek inzoomen,...).  Een andere optie is
"Statisch", waarbij de figuren vast zijn.}

\item{PathWD}{Het path van de working directory, dus het path waarin het
validatierapport opgeslagen moet worden (default wordt het in de op dat
moment actieve working directory opgeslagen).}
}
\value{
De functie genereert een validatierapport (\code{.html}-bestand) in de
working directory met informatie en grafieken van de te controleren metingen.
De afwijkende metingen zijn in rood aangeduid (zie \code{?validatierapport} of
vignet voor meer informatie).

De functie geeft een dataframe terug met de te controleren metingen, met
behalve de informatie uit de databank een aantal berekende waarden:
\itemize{
\item{\code{H_D_finaal}: een geschatte hoogte voor de omtrekklasse volgens het
domeinmodel}
\item{\code{rsmeD}: de foutenschatting voor het domeinmodel}
\item{\code{H_VL_finaal}: een geschatte hoogte voor de omtrekklasse volgens het
Vlaams model waarvan het domeinmodel afgeleid is}
\item{\code{rmseVL}: de foutenschatting voor dit Vlaams model}
\item{\code{HogeRmse}: \code{TRUE} als het domeinmodel een hoge RMSE heeft, anders
\code{NA}}
}
}
\description{
Functie die de validatie uitvoert op het verschoven Vlaams model en een
overzicht geeft van de afwijkende metingen (zodat de gebruiker deze kan
valideren).

\code{validatie.afgeleid()} roept meerdere hulpfuncties op:
\itemize{
\item{\code{rmse.basis()} en \code{rmse.verschuiving()}}
\item{\code{afwijkendeMetingen()}}
\item{\code{validatierapport()}}
}

Voorafgaand aan het uitvoeren van deze laatste functie worden eerst de
slechtste modellen opgelijst (op basis van RMSE en afwijkende metingen).
}
