% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afwijkendeCurves.R
\name{afwijkendeCurves}
\alias{afwijkendeCurves}
\title{Geeft de afwijkende metingen uit een gegeven model}
\usage{
afwijkendeCurves(Basismodel, Data = NULL)
}
\arguments{
\item{Basismodel}{model per boomsoort}

\item{Data}{dataset op basis waarvan het model berekend is (nodig voor extra model)}
}
\value{
Dataframe met te controleren curves (incl. aantal metingen per curve)
}
\description{
Functie die de curvekarakteristieken van de modellen berekent (roept hiervoor de functie curvekarakteristieken op) en op basis van deze parameters de mogelijk slechtste modellen oplijst met de reden.
}
\details{
Deze functie kan ook gebruikt worden voor extra modellen
}
