% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afwijkendeMetingen.R
\name{afwijkendeMetingen}
\alias{afwijkendeMetingen}
\title{Geeft de afwijkende metingen uit een gegeven model}
\usage{
afwijkendeMetingen(Dataset)
}
\arguments{
\item{Dataset}{Dataframe met meetresultaten, geschatte waarden voor het domeinmodel en het Vlaamse model en de rmse voor het domeinmodel.}
}
\value{
lijst met afwijkende metingen (> 2,5 * rmse), inclusief vlag uit databank
}
\description{
Berekent afwijkende metingen, dit zijn metingen met een afwijking > 2,5 * rmse, en geeft deze weer volgens dalende afwijking
}

