% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afwijkendeMetingen.R
\name{afwijkendeMetingen}
\alias{afwijkendeMetingen}
\title{Geeft de afwijkende metingen uit een gegeven model}
\usage{
afwijkendeMetingen(Dataset, AantalDomHogeRMSE = 20)
}
\arguments{
\item{Dataset}{Dataframe met meetresultaten, geschatte waarden voor het
domeinmodel en het Vlaamse model en de rmse voor het domeinmodel.  Deze
dataframe moet dezelfde velden bevatten als de dataframes in de list
teruggegeven in de functie `initiatie()`, en daarnaast de velden `H_D_finaal`,
`rmseD` en `maxResid`.}

\item{AantalDomHogeRMSE}{Standaard worden de 20 domeinen met de hoogste `RMSE`
geselecteerd en voor elk van deze domeinen worden minstens 10 afwijkende
metingen geselecteerd. `AantalDomHogeRMSE` laat toe om dit aantal van 20
domeinen aan te passen.}
}
\value{
Lijst met afwijkende metingen (> 2,5 * rmse), inclusief vlag uit
databank
}
\description{
Berekent afwijkende metingen, dit zijn metingen met een afwijking > 2,5 *
rmse, en geeft deze weer volgens dalende afwijking.  Voor de 20 domeinen met
de hoogste `RMSE`, waarbij de "foutenmarge" breder is, worden minstens 10
meetresultaten geselecteerd als "afwijkend" (nl. de 10 meetresultaten met de
hoogste error).
}
