% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvekarakteristieken.R
\name{curvekarakteristieken}
\alias{curvekarakteristieken}
\title{Berekent curvekarakteristieken van de domeinmodellen}
\usage{
curvekarakteristieken(Basismodel, Data = NULL)
}
\arguments{
\item{Basismodel}{model per boomsoort}

\item{Data}{meetgegevens (enkel nodig voor model per boomsoort-domein-combinatie)}
}
\value{
dataframe met curvekarakteristieken per domein en boomsoort
}
\description{
Berekent curvekarakteristieken die toelaten om mogelijke afwijkingen in curvevorm te detecteren: extremen (min en max), buigpunt,...
}
