% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhcurvesrapport.R
\name{dhcurvesrapport}
\alias{dhcurvesrapport}
\title{Stelt de finale dhcurves grafisch voor in een rapport}
\usage{
dhcurvesrapport(
  OutputIVANHO,
  Datalijst,
  Bestandsnaam = "dhcurves.html",
  verbose = TRUE,
  PathWD = getwd()
)
}
\arguments{
\item{OutputIVANHO}{Lijst met geschatte hoogte per omtrekklasse (dataframe
zoals gegenereerd door de functie `outputIVANHO()`).}

\item{Datalijst}{Datasets met gemeten waarden die gebruikt zijn voor de
modellen, zoals gegenereerd door de functie `initiatie()`.}

\item{Bestandsnaam}{Een naam voor het rapport (html-bestand) dat gegenereerd
wordt, bestaande uit een string die eindigt op `.html`}

\item{verbose}{Dit geeft de toestand van het systeem aan en zorgt ervoor dat
boodschappen niet onnodig gegeven worden.  (Default-waarde behouden.)}

\item{PathWD}{Het path van de working directory, dus het path waarin het
validatierapport opgeslagen moet worden (default wordt het in de op dat
moment actieve working directory opgeslagen).}
}
\value{
De functie genereert in de working directory (of opgegeven directory)
een rapport (html) met de te controleren modellen.  Hierin wordt per model
(boomsoort-domeincombinatie) de volgende algemene informatie vermeld:
boomsoort, domein (+ ID), modeltype, aantal metingen, rmse en bruikbaar
interval.

Daaronder wordt telkens grafisch volgende info weergegeven:
\itemize{
  \item{een puntenwolk die de metingen voorstelt (geen individuele metingen,
    maar een jitter)}
  \item{curve van het domeinmodel, aangeduid met zwarte ruiten}
  \item{grenzen van het bruikbaar interval (curves eindigen bij de
    klassemiddens die overeenkomen met deze grenzen)}
}
}
\description{
De functie genereert een rapport (html-bestand) in de working directory
(of opgegeven directory) met informatie en grafieken van de finale
modellen.

Deze functie rendert het bestand `dhcurverapport.Rmd`, dat voor elke
boomsoort-domeincombinatie child `dhcurve.Rmd` toevoegt.
}
