% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_afgeleid.R
\name{fit.afgeleid}
\alias{fit.afgeleid}
\title{Afgeleid model (= verschuiving t.o.v. Vlaams model) fitten op basis van het basismodel en de opgegeven dataset}
\usage{
fit.afgeleid(Data.afgeleid, Basismodel)
}
\arguments{
\item{Data.afgeleid}{dataframe 10-50}

\item{Basismodel}{model per boomsoort}
}
\value{
verschuiving per boomsoort en domein (dit is geen nieuw model, wel een verschuiving t.o.v. het Vlaams model)
}
\description{
Functie die het Vlaams model verschuift op basis van domeinen met 10 - 50 bomen (Domeinmodel = Vlaams model - residu’s): op basis van omtrekklassen > 0.5 m
}

