% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_afgeleid.R
\name{fit.afgeleid}
\alias{fit.afgeleid}
\title{Afgeleid model (= verschuiving t.o.v. Vlaams model) fitten op basis van het basismodel en de opgegeven dataset}
\usage{
fit.afgeleid(Data.afgeleid, Basismodel)
}
\arguments{
\item{Data.afgeleid}{Data.afgeleid dataframe 10-50}

\item{Basismodel}{model per boomsoort}
}
\value{
list met model per boomsoort-domein-combinatie
}
\description{
Functie die een model berekent voor domeinen met 10 - 50 bomen (binnen omtrekklassen > 0.5 m) door verschuiving van het Vlaams model voor die boomsoort.
}
