% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_afgeleid.R
\name{fit.afgeleid}
\alias{fit.afgeleid}
\title{Fit afgeleid model op basis van het basismodel en de opgegeven dataset}
\usage{
fit.afgeleid(Data.afgeleid, Basismodel)
}
\arguments{
\item{Data.afgeleid}{De dataframe "Afgeleid" uit de list die teruggegeven
wordt bij de functie \code{initiatie()}.  (Een dataframe met minimaal de velden
\code{BMS} (boomsoort), \code{IDbms} (identificatienummer van de boomsoort),
\code{DOMEIN_ID} (identificatienummer van het domein), \code{BOS_BHI} (domeinnaam),
\code{HOOGTE}, \code{Omtrek},
\code{logOmtrek}, \code{logOmtrek2}, \code{Status} (van data: "niet gecontroleerd",
"gecontroleerd",...), \code{nBomen} (aantal bomen per boomsoort-domeincombinatie),
\code{nBomenInterval} (aantal bomen per boomsoort-domeincombinatie binnen
bruikbaar interval), \code{nBomenIntervalOmtrek05} (aantal bomen per
boomsoort-domeincombinatie binnen bruikbaar interval met omtrek > 0.5 m),
\code{Q5k} en \code{Q95k} ("bruikbaar interval").)}

\item{Basismodel}{Model per boomsoort zoals teruggegeven door de functie
\code{fit.basis()}: tibble met de velden \code{BMS} (boomsoort) en \code{Model}
(\code{lme}-object met het gefit mixed model voor die boomsoort)}
}
\value{
List met 2 tibbles:
\itemize{
\item{tibble met velden \code{BMS} (boomsoort), \code{DOMEIN_ID} en \code{Model}
(\code{lm}-object van model per boomsoort-domeincombinatie)}
\item{tibble met de ingevoerde dataset, waaraan het veld \code{H_VL_finaal} (de
hoogteschatting volgens het Vlaams model) toegevoegd is}
}
}
\description{
Functie die op basis van de opgegeven dataset een model opstelt per
boomsoort-domeincombinatie door verschuiving van het Vlaams model voor die
boomsoort (afgeleid uit het basismodel). Dit komt overeen met het fitten van
het lineair model \eqn{Hoogte_{afgeleid} \sim \delta A + Hoogte_{Vlaams}}{%
Hoogte(afgeleid) ~ \delta A + Hoogte(Vlaams)} op basis van de opgegeven
dataset, waarbij \eqn{Hoogte_{Vlaams}}{Hoogte(Vlaams)} de hoogteschatting
is volgens het
Vlaams model.
}
