% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_basis.R
\name{fit.basis}
\alias{fit.basis}
\title{Fit basismodel op basis van de opgegeven dataset}
\usage{
fit.basis(Data.basis)
}
\arguments{
\item{Data.basis}{Dataframe met minimaal de velden BMS (boomsoort), DOMEIN_ID (identificatienummer van het domein), HOOGTE, logOmtrek en logOmtrek2.  Om alle verdere stappen van de analyse te kunnen doorlopen, wordt best de dataframe "Basis" gebruikt uit de list die teruggegeven wordt bij de functie initiatie.}
}
\value{
Dataframe (tibble) met de velden BMS (boomsoort) en Model (lme-object met het gefit mixed model voor die boomsoort)
}
\description{
Functie die het lineair mixed model \eqn{Hoogte_{basis} \sim A_{Vlaams} + A_{domein} + (B_{Vlaams} + B_{domein}) \log({Omtrek}) + (C_{Vlaams} + C_{domein}) \log(Omtrek)^2}{Hoogte(basis) ~ A(Vlaams) + A(domein) + (B(Vlaams) + B(domein)).log(Omtrek) + (C(Vlaams) + C(domein)).log(Omtrek)^2} fit op basis van de opgegeven dataset.
}
