% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lokaal.R
\name{fit.lokaal}
\alias{fit.lokaal}
\title{Lokaal model fitten op basis van de opgegeven dataset}
\usage{
fit.lokaal(Data.lokaal)
}
\arguments{
\item{Data.lokaal}{De dataframe \code{Lokaal} uit de list die teruggegeven wordt
bij de functie \code{initiatie()}.  (Een dataframe met minimaal de velden \code{BMS}
(boomsoort), \code{IDbms} (identificatienummer van de boomsoort), \code{DOMEIN_ID}
(identificatienummer van het domein), \code{BOS_BHI} (domeinnaam), \code{HOOGTE},
\code{Omtrek}, \code{logOmtrek}, \code{logOmtrek2},
\code{Status} (van data: "niet gecontroleerd", "gecontroleerd",...),
\code{nBomen} (aantal bomen per boomsoort-domeincombinatie),
\code{nBomenInterval} (aantal bomen per boomsoort-domeincombinatie binnen
bruikbaar interval), \code{nBomenOmtrek05} (aantal bomen per
boomsoort-domeincombinatie binnen bruikbaar interval met omtrek > 0.5 m),
\code{Q5k} en \code{Q95k} ("bruikbaar interval").)}
}
\value{
Dataframe (tibble) met de velden \code{BMS} (boomsoort), \code{DOMEIN_ID} en
\code{Model} (\code{lm}-object met het gefit lineair model voor die
boomsoort-domeincombinatie).
}
\description{
Functie die het lineair (fixed) model
\eqn{Hoogte_{lokaal} \sim A_{domein} + B_{domein} \log({Omtrek}) +
C_{domein} \log(Omtrek)^2}{Hoogte(lokaal) ~ A(domein) + B(domein).log(Omtrek)
+ C(domein).log(Omtrek)^2} fit op basis van de opgegeven dataset.
}
