% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lokaal.R
\name{fit.lokaal}
\alias{fit.lokaal}
\title{Lokaal model fitten op basis van de opgegeven dataset}
\usage{
fit.lokaal(Data.lokaal)
}
\arguments{
\item{Data.lokaal}{Dataframe met minimaal de velden BMS (boomsoort), DOMEIN_ID (identificatienummer van het domein), HOOGTE, logOmtrek en logOmtrek2.  Om alle verdere stappen van de analyse te kunnen doorlopen, wordt best de dataframe "Lokaal" gebruikt uit de list die teruggegeven wordt bij de functie initiatie.}
}
\value{
Dataframe (tibble) met de velden BMS (boomsoort), DOMEIN_ID en Model (lm-object met het gefit lineair model voor die boomsoort-domeincombinatie).
}
\description{
Functie die het lineair (fixed) model \eqn{Hoogte_{lokaal} \sim A_{domein} + B_{domein} \log({Omtrek}) + C_{domein} \log(Omtrek)^2}{Hoogte(lokaal) ~ A(domein) + B(domein).log(Omtrek) + C(domein).log(Omtrek)^2} fit op basis van de opgegeven dataset.
}
