% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoogteschatting_afgeleid.R
\name{hoogteschatting.afgeleid}
\alias{hoogteschatting.afgeleid}
\title{Hoogteschatting op basis van opgegeven afgeleid model}
\usage{
hoogteschatting.afgeleid(Afgeleidmodel, DataAfgeleidmodel)
}
\arguments{
\item{Afgeleidmodel}{verschoven Vlaams model voor 1 boomsoort-domein-combinatie}

\item{DataAfgeleidmodel}{de gegevens die hierbij horen: meetresultaten en het berekend Vlaams model (Deze dataset is het tweede item van de list die teruggegeven wordt bij de functie fit.afgeleid)}
}
\value{
dataframe met de meetresultaten en de schattingen van de hoogtes voor het domeinmodel en de Vlaamse model
}
\description{
Functie die de gemiddelde hoogte per omtrekklasse schat voor de domeincurves en Vlaamse curves van het opgegeven afgeleid model.  De teruggegeven dataframe kan gebruikt worden om grafieken te maken of afwijkende metingen te bestuderen.
}
