% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoogteschatting_afgeleid.R
\name{hoogteschatting.afgeleid}
\alias{hoogteschatting.afgeleid}
\title{Hoogteschatting op basis van opgegeven afgeleid model}
\usage{
hoogteschatting.afgeleid(Domeinsoortmodel, Domeinsoortdata)
}
\arguments{
\item{Domeinsoortmodel}{verschoven Vlaams model voor 1
boomsoort-domeincombinatie}

\item{Domeinsoortdata}{de gegevens die hierbij horen: meetresultaten voor 1
boomsoort-domeincombinatie}
}
\value{
dataframe met de meetresultaten en de schattingen van de hoogtes
voor het domeinmodel en de Vlaamse model
}
\description{
Functie die de gemiddelde hoogte per omtrekklasse schat voor de domeincurves
en Vlaamse curves van het opgegeven afgeleid model.  De teruggegeven
dataframe kan gebruikt worden om grafieken te maken of afwijkende metingen
te bestuderen.  Opgelet!  In tegenstelling tot de meeste functies van dit
package werkt deze functie op basis van 1 model en de bijhorende
meetgegevens.  Zie voorbeeld voor een methode om deze functie te kunnen
toepassen op de volledige dataset \code{Afgeleidmodel}.
}
\examples{
library(dplyr)

#Datasets inladen en het basismodel en afgeleid model berekenen
Data <- testdataset()
Datalijst <- initiatie(Data)
Data.basis <- Datalijst[["Basis"]]
Basismodel <- fit.basis(Data.basis)
Data.afgeleid <- Datalijst[["Afgeleid"]]
Afgeleidmodel <- fit.afgeleid(Data.afgeleid, Basismodel)

#De hoogteschatting voor een afgeleid model
Afgeleidmodel[[1]] \%>\%
  inner_join(
    Afgeleidmodel[[2]],
    by = c("BMS", "DOMEIN_ID")
  ) \%>\%
  group_by(
    .data$BMS,
    .data$DOMEIN_ID
  ) \%>\%
  do(
    hoogteschatting.afgeleid(.$Model[[1]],
                              select(., -Model))
  ) \%>\%
  ungroup()

}
