% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiatie.R
\name{initiatie}
\alias{initiatie}
\title{Initiatiestap met opsplitsen van dataset op basis van aantal metingen per domein}
\usage{
initiatie(Data, Uitzonderingen = data.frame(DOMEIN_ID = "", BMS = "",
  min_basis = NA, min_afgeleid = NA, stringsAsFactors = FALSE),
  Bestandsnaam = "VerwijderdeGegevens.html", verbose = TRUE,
  PathWD = getwd(), min_basismodel = 50, min_domeinen_basismodel = 6,
  min_afgeleidmodel = 10)
}
\arguments{
\item{Data}{dataframe met alle metingen waarop het model gebaseerd moet zijn (m.u.v. afgekeurde of te negeren metingen).  Velden DOMEIN_ID, BOS_BHI, IDbms, BMS, C13, HOOGTE  evt. TYPE_METING en JAAR, die worden bij rmse.basis als groeperende variabelen gebruikt.  C13 moet in centimeter opgegeven worden (maar wordt omgezet naar meter om de berekeningen uit te voern) en HOOGTE in meter.}

\item{Uitzonderingen}{lijst met uitzonderingen op min. 50 en min. 10 bomen.  Velden DOMEIN_ID, BMS, min_basis (= vervangende waarde voor 50), min_afgeleid (= vervangende waarde voor 10)}

\item{Bestandsnaam}{Een naam voor het html-bestand dat gegenereerd wordt, bestaande uit een string die eindigt op '.html'}

\item{verbose}{geeft de toestand van het systeem aan, om te zorgen dat boodschappen niet onnodig gegeven worden}

\item{PathWD}{Het path van de working directory, dus het path waarin het validatierapport opgeslagen moet worden (default de working directory)}

\item{min_basismodel}{tijdelijk toegevoegd voor testen}

\item{min_domeinen_basismodel}{tijdelijk toegevoegd voor testen}

\item{min_afgeleidmodel}{tijdelijk toegevoegd voor testen}
}
\value{
Een list van dataframes:

- dataframe > 50 metingen en min. 6 domeinen

- dataframe 10-50

- dataframe > 50 metingen en < 6 domeinen

- dataframe met metingen van domeinen en boomsoorten waar geen model voor opgesteld kan worden

En een validatierapport waarin een overzicht gegeven wordt van de verwijderde gegevens (omtrek > 2.4 m of < 0.2 m)
}
\description{
Deze functie bevat als initiële stap het berekenen van een aantal extra variabelen (bruikbaar interval (Q5 en Q95), Omtrek, logOmtrek, logOmtrek2, aantal metingen van bruikbaar interval en van interval > 0.5 m).  Bomen met omtrek > 2,3 m worden niet meegenomen voor de analyses en dus verwijderd uit de dataset.
}
\details{
Daarna splitst de functie de data op in 3 delen die voor de verschillende analyses kunnen gebruikt worden:

(1) boomsoorten waarvoor op minimum 6 domeinen veel metingen uitgevoerd zijn (> 50 metingen), op basis waarvan betrouwbare domeinmodellen en een betrouwbaar Vlaams model berekend kan worden (= basismodel),

(2) domeinen met minder metingen (10 - 50 metingen) van boomsoorten waarvoor een Vlaams model berekend kan worden (dus boomsoorten die in dataset (1) voorkomen), op basis waarvan het Vlaams model verschoven kan worden om een domeinspecifiek te model te bekomen (= afgeleid model),

(3) domeinen met veel metingen voor een boomsoort (> 50 metingen) waarvan er te weinig domeinen (< 6) zijn met voldoende metingen om een Vlaams model op te stellen.  Voor deze boomsoort-domein-combinaties kan een domeinspecifiek model opgesteld worden (maar geen Vlaams model voor de boomsoort, dus voor domeinen met < 50 metingen kan hier geen model gemaakt worden)(= lokaal model), en

(4) metingen van de domein-boomsoort-combinaties die niet tot de 3 voorgaande categorieën behoren en waar dus geen model voor opgesteld kan worden.

De grenswaarden 50 en 10 zijn gebaseerd op het aantal metingen binnen het interval 0,5 - 2,3 m en binnen het bruikbaar interval.  Bij de opsplitsing worden de data meteen gecleand, waarbij metingen met omtrek > 2,4 m en metingen buiten het bruikbaar interval sowieso weggelaten worden; voor het afgeleid model worden ook de metingen met omtrek <= 0,5 m weggelaten.
}
