% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoercontrole.R
\name{invoercontrole}
\alias{invoercontrole}
\title{Hulpfunctie die ingevoerd object controleert}
\usage{
invoercontrole(Data, Type)
}
\arguments{
\item{Data}{Het object dat moet gecontroleerd worden (dataframe of list)}

\item{Type}{Het type controle dat moet gebeuren, verwijzend naar de
functie(s) waarvoor het gebruikt wordt.  Mogelijke waarden zijn:
'initiatie', 'fit', 'basismodel', 'lokaalmodel', 'afgeleidmodel' en
'afgeleidedata'}
}
\value{
Een foutmelding of de geteste dataframe.
}
\description{
Omdat de controle op de invoer voor meerdere functies gelijkaardig is, is
deze interne functie ontwikkeld die voor de vaak voorkomende parameters van
de functies uit dhcurve controleert of de invoer voldoet aan de vereisten.
}
