% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputIVANHO.R
\name{outputIVANHO}
\alias{outputIVANHO}
\title{Berekent de geschatte hoogtes per omtrekklasse}
\usage{
outputIVANHO(
  Basismodel = NULL,
  Afgeleidmodel = NULL,
  Lokaalmodel = NULL,
  Data.lokaal = NULL
)
}
\arguments{
\item{Basismodel}{Model per boomsoort zoals teruggegeven door de functie
\code{fit.basis()}: tibble met de velden \code{BMS} (boomsoort) en \code{Model}
(\code{lme}-object met het gefit mixed model voor die boomsoort).}

\item{Afgeleidmodel}{Model per domein-boomsoortcombinatie zoals teruggegeven
door de functie \code{fit.afgeleid()}: list met 2 tibbles.}

\item{Lokaalmodel}{Model per boomsoort-domeincombinatie zoals teruggegeven
door de functie \code{fit.lokaal()}: tibble met de velden \code{BMS} (boomsoort),
\code{DOMEIN_ID} en \code{Model} (\code{lm}-object met het gefit lineair model voor die
boomsoort-domeincombinatie).}

\item{Data.lokaal}{Dataset op basis waarvan het opgegeven lokaal model
berekend is.}
}
\value{
Dataframe met geschatte hoogtes per domein en per boomsoort met
velden:
\itemize{
\item{\code{DOMEIN_ID}: domeincode}
\item{\code{BOS_BHI}: domeinnaam}
\item{\code{IDbms} (identificatienummer van de boomsoort)}
\item{\code{BMS}: boomsoort}
\item{\code{Modeltype} ("basismodel"(= eigen model op basis van mixed model) /
"afgeleid model"(= verschoven Vlaams model, afgeleid van fixed factor uit
basismodel) /
"Vlaams model"(= fixed factor uit basismodel, niet toegevoegd
omdat niet relevant) / "lokaal model"(= apart model voor 1
boomsoort-domein-combinatie) / "geen model"(= boomsoort-domein-combinatie
waarvoor geen model berekend kan worden))}
\item{\code{Omtrek}: klassenmidden van omtrekklasse}
\item{\code{OmtrekklassetypeID}: het overeenkomstige identificatienummer van de
omtrekklasse}
\item{\code{Omtrekklasse}}
\item{\code{Hoogte}: de geschatte hoogte}
\item{\code{RMSE} (root mean square error, zie vignette voor meer info:
\code{vignette("Handleiding", package = "dhcurve")})}
}
}
\description{
De functie \code{outputIVANHO()} voert de berekeningen uit die leiden tot de
gevraagde tabel om in IVANHO te importeren: een tabel met per boomsoort,
domein en omtrekklasse een schatting van de hoogte.  Als de curve een
maximum hoogte vertoont binnen het bestudeerde interval, wordt deze
maximumwaarde als hoogte meegegeven aan alle omtrekklassen hoger dan de
omtrekklasse van dit maximum.  (Dus verschillend van de validatierapporten
daalt de hoogte hier niet terug na het maximum, maar de hoogste waarde wordt
aangehouden.)

Voor deze functie worden volgende hulpfuncties aangeroepen:
\code{hoogteschatting.basis()}, \code{hoogteschatting.afgeleid()} en
\code{curvekarakteristieken()}.
}
