% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultaat.R
\name{resultaat}
\alias{resultaat}
\title{berekent de modelparameters per domein}
\usage{
resultaat(Basismodel, Afgeleidmodel, Lokaalmodel, Data.lokaal,
  Data.onbruikbaar = NULL)
}
\arguments{
\item{Basismodel}{model per boomsoort}

\item{Afgeleidmodel}{verschuiving per boomsoort en domein (Vlaams model)}

\item{Lokaalmodel}{model per boomsoort-domein-combinatie}

\item{Data.lokaal}{data voor model per boomsoort-domein-combinatie}

\item{Data.onbruikbaar}{evt. lijst met meetresultaten van domein-boomsoort-combinaties waarvoor geen model opgesteld kan worden}
}
\value{
dataframe met modellen per domein en per boomsoort met velden:

- ModelID

- DomeinID

- BoomsoortID

- Modeltype ('basismodel'(= eigen model op basis van mixed model) / ‘afgeleid model'(= verschoven Vlaams model, afgeleid van fixed factor uit basismodel) / ‘Vlaams model’(= fixed factor uit basismodel, niet toegevoegd omdat niet relevant) / 'lokaal model'(= apart model voor 1 boomsoort-domein-combinatie) / 'geen model'(= boomsoort-domein-combinatie waarvoor geen model berekend kan worden))

- paramaters A,B en C

- bruikbaar interval

- rmse

- aantal metingen waarop model gebaseerd is

- aantal metingen > 0.5 m (dus waarop rmse-berekening gebaseerd is)

evt. kan een tweede dataframe toegevoegd worden met Vlaamse modellen per boomsoort, of deze kan toegevoegd worden aan de vorige dataframe, waarbij DomeinID leeg gelaten wordt of een specifieke waarde ‘Vlaams model’ krijgt
}
\description{
Berekeningen die leiden tot de gevraagde tabel.  Hiervoor worden volgende hulpfuncties aangeroepen:
}
\details{
modelparameters
rsme.basis
fit.afgeleid

Verder moeten parameters A, B en C uit het model gehaald worden (coef) en een aantal berekende gegevens uit de eerste stap toegevoegd worden.
}
