% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultaat.R
\name{resultaat}
\alias{resultaat}
\title{Berekent de modelparameters per domein}
\usage{
resultaat(
  Basismodel = NULL,
  Afgeleidmodel = NULL,
  Lokaalmodel = NULL,
  Data.lokaal = NULL,
  Data.onbruikbaar = NULL
)
}
\arguments{
\item{Basismodel}{Model per boomsoort zoals teruggegeven door de functie
\code{fit.basis()}: tibble met de velden \code{BMS} (boomsoort) en \code{Model}
(\code{lme}-object met het gefit mixed model voor die boomsoort).}

\item{Afgeleidmodel}{Model per boomsoort-domeincombinatie zoals teruggegeven
door de functie \code{fit.afgeleid()}: list met 2 tibbles.}

\item{Lokaalmodel}{Model per boomsoort-domeincombinatie zoals teruggegeven
door de functie \code{fit.lokaal()}: tibble met de velden \code{BMS} (boomsoort),
\code{DOMEIN_ID} en \code{Model} (\code{lm}-object met het gefit lineair model voor die
boomsoort-domeincombinatie).}

\item{Data.lokaal}{Dataset op basis waarvan het opgegeven lokaal model
berekend is.}

\item{Data.onbruikbaar}{Evt. lijst met meetresultaten van
boomsoort-domeincombinaties waarvoor geen model opgesteld kan worden, in de
vorm van een dataframe zoals de dataframe \code{Rest} uit de list die door de
functie \code{initiatie()} teruggegeven wordt.}
}
\value{
Dataframe met modellen per domein en per boomsoort met velden:
\itemize{
\item{\code{DomeinID} (identificatienummer voor domein)}
\item{\code{BMS} (boomsoort)}
\item{\code{Modeltype} ("basismodel"(= eigen model op basis van mixed model) /
"afgeleid model"(= verschoven Vlaams model, afgeleid van fixed factor uit
basismodel) /
"Vlaams model"(= fixed factor uit basismodel, niet toegevoegd
omdat niet relevant) / "lokaal model"(= eigen model voor 1
boomsoort-domeincombinatie) / "geen model"(= boomsoort-domeincombinatie
waarvoor minstens 1 boom opgemeten is maar geen model berekend kan
worden))}
\item{parameters \code{A}, \code{B} en \code{C} (zie 'Description' bovenaan)}
\item{bruikbaar interval (\code{Q5k} en \code{Q95k}, zie vignet voor meer info)}
\item{\code{RMSE} (root mean square error, zie vignet voor meer info)}
\item{\code{nBomen} (totaal aantal opgemeten bomen met omtrek tussen 0,2 en
2,4 m)}
\item{\code{nBomenInterval} (aantal metingen waarop model gebaseerd is)}
\item{\code{nBomenIntervalOmtrek05} (aantal metingen > 0.5 m, dus waarop
RMSE-berekening gebaseerd is)}
}

evt. kan een tweede dataframe toegevoegd worden met Vlaamse modellen per
boomsoort, of deze kan toegevoegd worden aan de vorige dataframe, waarbij
\code{DomeinID} leeg gelaten wordt of een specifieke waarde
"Vlaams model" krijgt
}
\description{
De functie resultaat berekent op basis van de opgegeven modellen voor elke
boomsoort-domeincombinatie de modelparameters A, B en C voor een model van
de vorm \eqn{Hoogte \sim A + B\log(Omtrek) + C\log(Omtrek)^2}{Hoogte ~ A +
B.log(Omtrek) + C.log(Omtrek)^2}.

Voor deze functie worden volgende hulpfuncties aangeroepen:
\code{modelparameters()}, \code{rsme.basis()} en \code{rsme.afgeleid()}

Verder worden parameters \code{A}, \code{B} en \code{C} uit het model gehaald en een
aantal in de functie \code{initiatie()} berekende gegevens toegevoegd.
}
