% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_afgeleid.R
\name{rmse.afgeleid}
\alias{rmse.afgeleid}
\title{RMSE berekenen van afgeleid model (verschoven Vlaams model)}
\usage{
rmse.afgeleid(Verschovenmodel, Boomsoort, Domein)
}
\arguments{
\item{Verschovenmodel}{afgeleid model per boomsoort-domein-combinatie}

\item{Boomsoort}{BMS}

\item{Domein}{DOMEIN_ID}
}
\value{
dataframe met rmse_verschuiving per boomsoort en domein
}
\description{
deze functie berekent de rmse op basis van verschil tussen geschatte domeinwaarde en gemeten waarde (voor omtrekklassen > 0.5 m)
}
\details{
Opgelet!  In tegenstelling tot de meeste functies van dit package werkt deze functie op basis van de meetgegevens van 1 model.  Zie voorbeeld voor een methode om deze functie te kunnen toepassen vertrekkend van het berekende model.
}
\examples{
library(dplyr)
#nog datasets toevoegen om deze voorbeelden te kunnen runnen
\dontrun{
Afgeleidmodel[[1]] \%>\%
  rowwise() \%>\%
  do_(
    ~rmse.afgeleid(.$Model, .$BMS, .$DOMEIN_ID)
  ) \%>\%
  ungroup()
#Nota: voor een berekening van de volledige RMSE van dit afgeleid model
#moet ook de RMSE van het basismodel in rekening gebracht worden
}

}
