% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_afgeleid.R
\name{rmse.afgeleid}
\alias{rmse.afgeleid}
\title{Berekent RMSE van afgeleid model}
\usage{
rmse.afgeleid(Verschovenmodel, Boomsoort, Domein)
}
\arguments{
\item{Verschovenmodel}{Afgeleid model voor één boomsoort-domein-combinatie (lm-object)}

\item{Boomsoort}{BMS}

\item{Domein}{DOMEIN_ID}
}
\value{
Dataframe met BMS(boomsoort), DOMEIN_ID, RmseVerschuiving
}
\description{
Deze functie berekent de rmse op basis van verschil tussen de hoogteschatting volgens het domeinmodel en de gemeten hoogte (voor omtrekklassen > 0.5 m).  Dit is slechts een deel van de totale rmse van het afgeleid model, want behalve deze rmse van de verschuiving moet er ook rekening gehouden met de rmse van het Vlaams model (basismodel).   Opgelet!  In tegenstelling tot de meeste functies van dit package werkt deze functie op basis van de meetgegevens van 1 model.  Zie voorbeeld voor een methode om deze functie te kunnen toepassen vertrekkend van het berekende model.
}
\examples{
library(dplyr)
#nog datasets toevoegen om deze voorbeelden te kunnen runnen
\dontrun{
Afgeleidmodel[[1]] \%>\%
  rowwise() \%>\%
  do_(
    ~rmse.afgeleid(.$Model, .$BMS, .$DOMEIN_ID)
  ) \%>\%
  ungroup()
#Nota: voor een berekening van de volledige RMSE van dit afgeleid model
#moet ook de RMSE van het basismodel in rekening gebracht worden
}

}
