% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_afgeleid.R
\name{rmse.afgeleid}
\alias{rmse.afgeleid}
\title{RMSE berekenen van afgeleid model (verschoven Vlaams model)}
\usage{
rmse.afgeleid(Verschovenmodel, Boomsoort, Domein)
}
\arguments{
\item{Verschovenmodel}{afgeleid model per boomsoort-domein-combinatie}

\item{Boomsoort}{BMS}

\item{Domein}{DOMEIN_ID}
}
\value{
dataframe met rmse_verschuiving per boomsoort en domein
}
\description{
deze functie berekent de rmse op basis van verschil tussen geschatte domeinwaarde en gemeten waarde (voor omtrekklassen > 0.5 m)
}
