% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_basis.R
\name{rmse.basis}
\alias{rmse.basis}
\title{RMSE berekenen van basismodel}
\usage{
rmse.basis(Basismodel)
}
\arguments{
\item{Basismodel}{model per boomsoort als argument meegeven en hier de nodige gegevens uit halen  (Vermits de 2 hoofdfuncties waarin deze hulpfunctie opgeroepen wordt allebei het argument model beschikbaar hebben en de dataframe niet, lijkt het me het meest logisch om hier van het model te vertrekken, dan moet het script om de meetgegevens uit het model te halen, enkel in deze functie geschreven worden)  Een alternatief is vertrekken van het dataframe > 50 en min. 6 domeinen}
}
\value{
dataframe met rmse_domein en rmse_Vlaams
}
\description{
Deze functie berekent de rmse op basis van testgroepen en omvat de volgende deelstappen:
}
\details{
- metingen opdelen in 6 testgroepen (veld testgroep)

- modellen fitten voor testgroepen, waarbij ze de functie fit.basis 6 keer oproept

- rmse berekenen voor domeinmodellen en Vlaams model op basis van testgroep-modellen
}

