% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_verschuiving.R
\name{rmse.verschuiving}
\alias{rmse.verschuiving}
\title{Berekent RMSE van afgeleid model}
\usage{
rmse.verschuiving(Verschovenmodel, Boomsoort, Domein)
}
\arguments{
\item{Verschovenmodel}{Afgeleid model voor één boomsoort-domein-combinatie
(lm-object)}

\item{Boomsoort}{BMS}

\item{Domein}{DOMEIN_ID}
}
\value{
Dataframe met BMS(boomsoort), DOMEIN_ID, RmseVerschuiving
}
\description{
Deze functie berekent de rmse op basis van verschil tussen de
hoogteschatting volgens het domeinmodel en de gemeten hoogte (voor
omtrekklassen > 0.5 m).  Dit is slechts een deel van de totale rmse van het
afgeleid model, want behalve deze rmse van de verschuiving moet hier ook de
rmse van het Vlaams model (basismodel) in rekening gebracht worden.
Opgelet!  In tegenstelling tot de meeste functies van dit package werkt deze
functie op basis van de meetgegevens van 1 model.  Zie voorbeeld voor een
methode om deze functie te kunnen toepassen vertrekkend van het berekende
model.
}
\examples{
library(dplyr)

#Datasets inladen en het basismodel en afgeleid model berekenen
Data <- testdataset()
Datalijst <- initiatie(Data)
Data.basis <- Datalijst[["Basis"]]
Basismodel <- fit.basis(Data.basis)
Data.afgeleid <- Datalijst[["Afgeleid"]]
Afgeleidmodel <- fit.afgeleid(Data.afgeleid, Basismodel)

# De rmse berekenen voor de verschuiving van het Vlaams model naar een
# afgeleid model
Afgeleidmodel[[1]] \%>\%
  rowwise() \%>\%
  do_(
    ~rmse.verschuiving(.$Model, .$BMS, .$DOMEIN_ID)
  ) \%>\%
  ungroup()
#Nota: voor een berekening van de volledige RMSE van een afgeleid model
#moet ook de RMSE van het basismodel in rekening gebracht worden

}
