% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdataset.R
\name{testdataset}
\alias{testdataset}
\title{Genereert testdataset}
\usage{
testdataset(
  nBomenDomein = c(200, 100, 500, 100, 100, 100, 50, 20, 8),
  BMS = "testboom",
  IDbms = 1,
  minOmtrek = 20,
  maxOmtrek = 239
)
}
\arguments{
\item{nBomenDomein}{Vector met aantal gemeten bomen per domein}

\item{BMS}{Naam van boomsoort}

\item{IDbms}{Identificatienummer van boomsoort}

\item{minOmtrek}{Minimale omtrek voor de data}

\item{maxOmtrek}{Maximale omtrek voor de data}
}
\value{
Dataframe met testdata voor 1 boomsoort met velden \code{BMS}, \code{IDbms},
\code{DOMEIN_ID}, \code{BOS_BHI}, \code{C13}, \code{HOOGTE}, \code{Status} en \code{ID}.
}
\description{
Deze functie genereert een testdataset die voldoet aan de opgegeven
vereisten.
De testdataset kan gebruikt worden om het package te testen, ze voldoet als
invoer voor functie \code{initiatie()}
}
