% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatie_basis.R
\name{validatie.basis}
\alias{validatie.basis}
\title{Validatie van het basismodel}
\usage{
validatie.basis(Basismodel, Data = NULL)
}
\arguments{
\item{Basismodel}{model per boomsoort}

\item{Data}{dataset op basis waarvan het model berekend is (nodig voor lokaal model)}
}
\value{
Dataframe met te controleren metingen en document (html/pdf) met te controleren curves (incl. aantal metingen per curve) en grafieken van te controleren metingen
}
\description{
Functie die alle nodige validaties uitvoert op het opgegeven (volledige) model en een overzicht geeft van de afwijkende metingen en slechte curves (zodat de gebruiker deze kan valideren). De functie roept meerdere hulpfuncties op:
}
\details{
- rmse.basis

- afwijkendeMetingen

- afwijkendeCurves

- validatierapport

Voorafgaand aan het uitvoeren van deze laatste functie worden eerst de slechtste modellen opgelijst (op basis van rmse, afwijkende metingen en afwijkende curves).

Deze functie kan ook gebruikt worden voor lokale modellen
}
