% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatie_uitbreiding.R
\name{validatie.uitbreiding}
\alias{validatie.uitbreiding}
\title{Validatie van een uitbreiding van een model naar hogere omtrekklassen}
\usage{
validatie.uitbreiding(
  Model,
  Dataset,
  AantalDomValidatie = 20,
  GoedgekeurdeUitbreidingen = NULL,
  Bestandsnaam = "Default",
  TypeRapport = "Dynamisch",
  PathWD = getwd()
)
}
\arguments{
\item{Model}{Basismodel per boomsoort of
lokaal model per boomsoort-domein-combinatie zoals teruggegeven
door de functies \code{fit.basis()} of \code{fit.lokaal()}}

\item{Dataset}{Dataset zoals teruggegeven door functie \code{initiatie()}, op
basis waarvan het opgegeven model (basismodel of lokaal model) berekend is.}

\item{AantalDomValidatie}{Standaard worden de 20 domeinen met de hoogste
gemiddelde afwijking (\code{DiffMediaan}) geselecteerd om te valideren.
(Hierbij worden \code{GoedgekeurdeUitbreidingen} niet meer getoond.)
\code{AantalDomHogeRMSE} laat toe om dit aantal van 20 domeinen aan te passen.}

\item{GoedgekeurdeUitbreidingen}{Optie om goedgekeurde uitbreidingen
niet meer te tonen in het validatierapport zolang er geen extra metingen
toegevoegd zijn voor de boomsoort-domeincombinatie (om als gebruiker enkel de
te keuren curves in het rapport over te houden). De goedgekeurde en dus te
negeren curves moeten opgegeven worden in een dataframe met velden
\code{DOMEIN_ID}, \code{BMS} en \code{nBomenTerugTonen}, met \code{nBomenTerugTonen} het aantal
bomen boven het bruikbaar interval (\code{nExtra}) in de
boomsoort-domeincombinatie vanaf wanneer de curve terug getoond moet worden.
(In dit geval wordt de curve uiteraard enkel
terug getoond als ze bij de 20 slechtste curves hoort, met hoogste
\code{diff_mediaan}.)}

\item{Bestandsnaam}{Een naam voor het validatierapport (\code{.html}-bestand) dat
gegenereerd wordt, bestaande uit een string die eindigt op \code{.html}}

\item{TypeRapport}{Default is "Dynamisch", waarbij de figuren in het
\code{.html}-bestand kunnen worden aangepast (meetgegevens weergeven door muis
erover te bewegen (inclusief \code{ID} als deze in de dataset meegegeven is),
items uit legende wegklikken, grafiek inzoomen,...).  Een andere optie is
"Statisch", waarbij de figuren vast zijn.}

\item{PathWD}{Het path van de working directory, dus het path waarin het
validatierapport opgeslagen moet worden (default wordt het in de op dat
moment actieve working directory opgeslagen).}
}
\value{
De functie genereert een validatierapport (\code{.html}-bestand) in de working
directory met informatie en grafieken van de te controleren modellen.
De weergegeven grafieken zijn gerangschikt van slechtste (= mediaan heeft
grootste afwijking) naar minder slecht.
Idee is om in dit rapport het blauwe deel van de grafiek te valideren.

De functie geeft een dataframe terug met per boomsoort-domeincombinatie de
omtrekklasse tot waar het model uitgebreid kan worden
(met velden \code{BMS}, \code{DOMEIN_ID} en \code{MaxOmtrek}).
Deze tabel kan als argument \code{Uitbreidingen} in functie
\code{outputIVANHO()} toegevoegd worden om deze uitbreidingen mee te nemen voor
het eindresultaat.
Belangrijk is hierbij dat afgekeurde curves eerst verwijderd zijn door
bij \code{initiatie()} via \code{Uitzonderingen} de variabele
\code{min_uitbreiden_model} op te trekken en daarna
\code{validatie.uitbreiding()} opnieuw gerund is met de aangepaste dataset.
}
\description{
Functie die toelaat om de opgegeven basismodellen of lokale modellen te
valideren ter hoogte van omtrekklassen boven het bruikbaar interval ingeval
hiervoor voldoende gegevens beschikbaar zijn.
De functie berekent de waarden \code{diff_min}, \code{diff_mediaan} en \code{diff_max} en
geeft een overzichtsrapport met de 20 slechtste niet eerder afgekeurde
curves (= curves met hoogste absolute waarde voor \code{diff_mediaan}),
waarvan het de bedoeling is dat de gebruiker ze valideert.
}
