% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatierapport.R
\name{validatierapport}
\alias{validatierapport}
\title{stelt de slechtste curves en afwijkende metingen grafisch voor in een rapport}
\usage{
validatierapport(SlechtsteModellen, AfwijkendeMetingen, Dataset,
  Bestandsnaam = "Validatie.html", verbose = TRUE)
}
\arguments{
\item{SlechtsteModellen}{lijst van de slechtste modellen, dit zijn modellen met hoge rmse, afwijkende vorm (op basis van extremen en buigpunten) en/of modellen met afwijkende metingen.  Deze bevat ook de info om weer te geven bij de grafieken (parameters)}

\item{AfwijkendeMetingen}{lijst met afwijkende metingen (zoals gegenereerd door de functie afwijkendeMetingen)}

\item{Dataset}{Dataset met gemeten waarden en geschatte waarde voor domeinmodel en Vlaams model (inclusief RMSE)}

\item{Bestandsnaam}{Een naam voor het html-bestand dat gegenereerd wordt, bestaande uit een string die eindigt op '.html'}

\item{verbose}{geeft de toestand van het systeem aan, om te zorgen dat boodschappen niet onnodig gegeven worden}
}
\value{
document (html/pdf) met te controleren curves (incl. aantal metingen per curve)
}
\description{
Deze functie rendert een Rmd-file.  In de Rmd-file wordt voor elk model uit de lijst een grafiek van dezelfde vorm gemaakt, op basis van aparte functie of child-rmd _maakGrafiek_ die voor elk item uit de lijst opnieuw opgeroepen wordt.
}
\details{
Op de grafieken moet volgende info getoond worden:

- puntenwolk metingen

- curve Vlaams model

- curve domeinmodel of verschoven Vlaams model (dus domeinspecifieke curve)

- extremen en buigpunten waar aanwezig (met vertikale stippellijn?)

- grenzen van bruikbaar interval

- vermelding van boomsoort, domein, aantal waarnemingen, rmse, evt. parameters A, B en C

- eventueel afwijkende metingen

Idee is dat in de grafieken een afwijkende kleur het probleem aangeeft: afwijkende meting, buigpunt,...  Daarnaast wordt het probleem ook tekstueel weergegeven
}
